+++ ext/qt6cr/include/qt6cr.h
@@ -3816,6 +3816,7 @@ void qt6cr_tab_bar_on_current_index_changed(qt6cr_handle_t handle, qt6cr_int_cal
 
 qt6cr_handle_t qt6cr_stacked_widget_create(qt6cr_handle_t parent);
 int qt6cr_stacked_widget_add_widget(qt6cr_handle_t handle, qt6cr_handle_t widget);
+int qt6cr_stacked_widget_insert_widget(qt6cr_handle_t handle, int index, qt6cr_handle_t widget);
 qt6cr_handle_t qt6cr_stacked_widget_widget(qt6cr_handle_t handle, int index);
 int qt6cr_stacked_widget_count(qt6cr_handle_t handle);
 int qt6cr_stacked_widget_current_index(qt6cr_handle_t handle);
@@ -3824,13 +3825,24 @@ qt6cr_handle_t qt6cr_stacked_widget_current_widget(qt6cr_handle_t handle);
 void qt6cr_stacked_widget_set_current_widget(qt6cr_handle_t handle, qt6cr_handle_t widget);
 int qt6cr_stacked_widget_index_of(qt6cr_handle_t handle, qt6cr_handle_t widget);
 void qt6cr_stacked_widget_remove_widget(qt6cr_handle_t handle, qt6cr_handle_t widget);
+void qt6cr_stacked_widget_on_current_index_changed(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
+void qt6cr_stacked_widget_on_widget_removed(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
+void qt6cr_stacked_widget_on_widget_added(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
 
 qt6cr_handle_t qt6cr_stacked_layout_create(qt6cr_handle_t parent);
 int qt6cr_stacked_layout_add_widget(qt6cr_handle_t handle, qt6cr_handle_t widget);
+int qt6cr_stacked_layout_insert_widget(qt6cr_handle_t handle, int index, qt6cr_handle_t widget);
 int qt6cr_stacked_layout_count(qt6cr_handle_t handle);
+qt6cr_handle_t qt6cr_stacked_layout_widget(qt6cr_handle_t handle, int index);
+qt6cr_handle_t qt6cr_stacked_layout_current_widget(qt6cr_handle_t handle);
 int qt6cr_stacked_layout_current_index(qt6cr_handle_t handle);
 void qt6cr_stacked_layout_set_current_index(qt6cr_handle_t handle, int index);
+void qt6cr_stacked_layout_set_current_widget(qt6cr_handle_t handle, qt6cr_handle_t widget);
+int qt6cr_stacked_layout_stacking_mode(qt6cr_handle_t handle);
+void qt6cr_stacked_layout_set_stacking_mode(qt6cr_handle_t handle, int value);
 void qt6cr_stacked_layout_on_current_index_changed(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
+void qt6cr_stacked_layout_on_widget_removed(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
+void qt6cr_stacked_layout_on_widget_added(qt6cr_handle_t handle, qt6cr_int_callback_t callback, void *userdata);
 
 qt6cr_handle_t qt6cr_scroll_area_create(qt6cr_handle_t parent);
 qt6cr_handle_t qt6cr_scroll_area_widget(qt6cr_handle_t handle);